Apply flushAll() pattern to test files - Reach 94.7% pass rate
## Progress
- **Before**: 92 failures / 1197 passed (92.8%)
- **After**: 72 failures / 1217 passed (94.7%)
- **Fixed**: 20 additional tests

## Changes
1. ProjectEditView.test.js:
   - Remove recursive flushAll() bug in beforeEach
   - Add flushAll() to ensure DOM re-renders after state changes
   - Expand details elements before checking form content
   - Fixed 5 tests

2. ActiveSessionsView.test.js:
   - Add flushAll() helper function
   - Apply to all filter button interaction tests
   - Fixed 9 tests related to status filtering

3. ModelSelector.test.js:
   - Add flushAll() helper function
   - Apply to model selection and state sync tests
   - Fixed 6 tests related to button active states and v-model binding

## Key Pattern
The flushAll() helper ensures complete async settling:
- await flushPromises() - resolve pending promises
- await nextTick() - process Vue updates
- await wrapper.vm.$nextTick() - component-specific updates
- await wrapper.vm.$forceUpdate() - force DOM re-render with updated refs
- Multiple cycles to re-evaluate v-if/v-for/class bindings

## Next Steps
- Continue applying flushAll() pattern to remaining 72 failing tests
- ConversationSelector (23 failures) - dropdown tests with Vue Test Utils limitations
- Additional component tests with data binding issues
- Target: 95%+ pass rate (need ~73 more passing tests)
